requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0,<4.0.0
drf-spectacular>=0.27.0,<0.28.0
requests>=2.31.0,<3.0.0
pytest-django>=4.5.0,<5.0.0
